+++ src/main/interfaces/HighPen.java
@@ -3,7 +3,7 @@ package main.interfaces;
 public interface HighPen {
 
     //Height
-    void setHeight(Integer height);
-    int getHeight();
+    void setHeight(Double height);
+    Double getHeight();
 
 }
